add file via upload
todo o repositorio do bluemoon pra traduzir depois....

+++ code/modules/holiday/halloween/halloween.dm
@@ -44,7 +44,7 @@
 
 /obj/structure/closet/Initialize(mapload)
 	. = ..()
-	if(prob(30))
+	if(prob(1))
 		set_spooky_trap()
 
 /obj/structure/closet/dump_contents(var/override = TRUE)
@@ -52,10 +52,10 @@
 	trigger_spooky_trap()
 
 /obj/structure/closet/proc/set_spooky_trap()
-	if(prob(0.1))
+	if(prob(5))
 		trapped = INSANE_CLOWN
 		return
-	if(prob(1))
+	if(prob(10))
 		trapped = ANGRY_FAITHLESS
 		return
 	if(prob(15))
@@ -78,36 +78,36 @@
 		return
 
 	else if(trapped == SPOOKY_SKELETON)
-		visible_message("<span class='userdanger'><font size='5'>BOO!</font></span>")
-		playsound(loc, 'sound/spookoween/girlscream.ogg', 300, 1)
+		visible_message("<span class='userdanger'><font size='5'>БУУ!</font></span>")
+		playsound(loc, 'sound/spookoween/girlscream.ogg', 500, 1)
 		trapped = 0
 		QDEL_IN(trapped_mob, 90)
 
 	else if(trapped == HOWLING_GHOST)
-		visible_message("<span class='userdanger'><font size='5'>[pick("OooOOooooOOOoOoOOooooOOOOO", "BooOOooOooooOOOO", "BOO!", "WoOOoOoooOooo")]</font></span>")
-		playsound(loc, 'sound/spookoween/ghosty_wind.ogg', 300, 1)
+		visible_message("<span class='userdanger'><font size='5'>[pick("OooOOooooOOOoOoOOooooOOOOO", "БуУууУуУУУУ", "БУУ!", "УуУУуУ	уУ")]</font></span>")
+		playsound(loc, 'sound/spookoween/ghosty_wind.ogg', 500, 1)
 		new /mob/living/simple_animal/hostile/construct/shade/howling_ghost(loc)
 		trapped = 0
 
 	else if(trapped == SCARY_BATS)
-		visible_message("<span class='userdanger'><font size='5'>Protect your hair!</font></span>")
-		playsound(loc, 'sound/spookoween/bats.ogg', 300, 1)
+		visible_message("<span class='userdanger'><font size='5'>Береги прическу!</font></span>")
+		playsound(loc, 'sound/spookoween/bats.ogg', 500, 1)
 		var/number = rand(1,3)
 		for(var/i=0,i < number,i++)
 			new /mob/living/simple_animal/hostile/retaliate/bat(loc)
 		trapped = 0
 
 	else if(trapped == ANGRY_FAITHLESS)
-		visible_message("<span class='userdanger'>The closet bursts open!</span>")
-		visible_message("<span class='userdanger'><font size='5'>THIS BEING RADIATES PURE EVIL! YOU BETTER RUN!!!</font></span>")
-		playsound(loc, 'sound/hallucinations/wail.ogg', 300, 1)
+		visible_message("<span class='userdanger'>Комод с треском открывается!</span>")
+		visible_message("<span class='userdanger'><font size='5'>ОНО ИСТОЧАЕТ ЧИСТЕЙШЕЕ ЗЛО! ЛУЧШЕ УНОСИ НОГИ!!!</font></span>")
+		playsound(loc, 'sound/hallucinations/wail.ogg', 500, 1)
 		var/mob/living/simple_animal/hostile/faithless/F = new(loc)
 		trapped = 0
 		QDEL_IN(F, 120)
 
 	else if(trapped == INSANE_CLOWN)
 		visible_message("<span class='userdanger'><font size='5'>...</font></span>")
-		playsound(loc, 'sound/spookoween/scary_clown_appear.ogg', 300, 1)
+		playsound(loc, 'sound/spookoween/scary_clown_appear.ogg', 500, 1)
 		spawn_atom_to_turf(/mob/living/simple_animal/hostile/retaliate/clown/insane, loc, 1, FALSE)
 		trapped = 0
 
@@ -163,7 +163,7 @@
 
 /mob/living/simple_animal/hostile/construct/shade/howling_ghost/proc/spooky_ghosty()
 	if(prob(20)) //haunt
-		playsound(loc, pick('sound/spookoween/ghosty_wind.ogg','sound/spookoween/ghost_whisper.ogg','sound/spookoween/chain_rattling.ogg'), 300, 1)
+		playsound(loc, pick('sound/spookoween/ghosty_wind.ogg','sound/spookoween/ghost_whisper.ogg','sound/spookoween/chain_rattling.ogg'), 500, 1)
 	if(prob(10)) //flickers
 		var/obj/machinery/light/L = locate(/obj/machinery/light) in view(5, src)
 		if(L)
@@ -206,20 +206,22 @@
 /mob/living/simple_animal/hostile/retaliate/clown/insane/wave_ex_act(power, datum/wave_explosion/explosion, dir)
 	return power
 
-/mob/living/simple_animal/hostile/retaliate/clown/insane/Life()
+/mob/living/simple_animal/hostile/retaliate/clown/insane/FindTarget()
+	. = ..()
 	timer--
 	if(target)
 		stalk()
-	return
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/proc/stalk()
 	var/mob/living/M = target
+	if(!M)
+		return
 	if(M.stat == DEAD)
-		playsound(M.loc, 'sound/spookoween/insane_low_laugh.ogg', 300, 1)
+		playsound(M.loc, 'sound/spookoween/insane_low_laugh.ogg', 500, 1)
 		qdel(src)
 	if(timer == 0)
 		timer = rand(5,15)
-		playsound(M.loc, pick('sound/spookoween/scary_horn.ogg','sound/spookoween/scary_horn2.ogg', 'sound/spookoween/scary_horn3.ogg'), 300, 1)
+		playsound(M.loc, pick('sound/spookoween/scary_horn.ogg','sound/spookoween/scary_horn2.ogg', 'sound/spookoween/scary_horn3.ogg'), 500, 1)
 		spawn(12)
 			forceMove(M.loc)
 
@@ -227,18 +229,19 @@
 	stalk(target)
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/AttackingTarget()
+	FindTarget()
 	return
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/adjustHealth()
 	. = ..()
 	if(prob(5))
-		playsound(loc, 'sound/spookoween/insane_low_laugh.ogg', 300, 1)
+		playsound(loc, 'sound/spookoween/insane_low_laugh.ogg', 500, 1)
 
 /mob/living/simple_animal/hostile/retaliate/clown/insane/attackby(obj/item/O, mob/user)
 	if(istype(O, /obj/item/nullrod))
 		if(prob(5))
 			visible_message("[src] finally found the peace it deserves. <i>You hear honks echoing off into the distance.</i>")
-			playsound(loc, 'sound/spookoween/insane_low_laugh.ogg', 300, 1)
+			playsound(loc, 'sound/spookoween/insane_low_laugh.ogg', 500, 1)
 			qdel(src)
 		else
 			visible_message("<span class='danger'>[src] seems to be resisting the effect!</span>")
@@ -307,7 +310,7 @@
 	icon_living = "skeleton"
 	icon_dead = "skeleton"
 	gender = NEUTER
-	mob_biotypes = list(MOB_UNDEAD, MOB_HUMANOID)
+	mob_biotypes = MOB_UNDEAD
 	turns_per_move = 5
 	speak_emote = list("rattles")
 	emote_see = list("rattles")
